🐛 fix: fix canonical url
